New Question type: Anagram Question
Asks for the anagram of a word from a list, for example:
  which of the following is an anagram of "listen": enlists, google, banana, inlets

The anagrams are all stored in a yaml file called anagrams.yaml. You specify the question word, the correct answer and a series of wrong answers like the example in the file.

I may have also removed all trailing whitespace from question_factory.rb